better useMutation types

+++ src/lib/react-query/query-mutations/user-media.mutations.ts
@@ -1,7 +1,7 @@
-import {Label} from "@/lib/components/types";
 import {useMutation, useQueryClient} from "@tanstack/react-query";
 import {LabelAction, MediaType, Status} from "@/lib/server/utils/enums";
 import {queryKeys} from "@/lib/react-query/query-options/query-options";
+import {HistoryOptionsType, Label, MediaDetailsOptionsType, MediaListOptionsType, ProfileOptionsType, UserMedia} from "@/lib/components/types";
 import {postAddMediaToList, postDeleteUserUpdates, postEditUserLabel, postRemoveMediaFromList, postUpdateUserMedia} from "@/lib/server/functions/user-media";
 
 
@@ -14,22 +14,28 @@ export type DetailsUserListKeys = ReturnType<(typeof detailsUserListKeys)[number
 
 export const useDeleteUpdatesMutation = (queryKey: DeleteUpdatesKeys) => {
     const queryClient = useQueryClient();
-    return useMutation<any, Error, { updateIds: number[], returnData?: boolean }>({
-        mutationFn: ({ updateIds, returnData = false }) => postDeleteUserUpdates({ data: { updateIds, returnData } }),
+    return useMutation({
+        mutationFn: ({ updateIds, returnData = false }: { updateIds: number[], returnData?: boolean }) => {
+            return postDeleteUserUpdates({ data: { updateIds, returnData } })
+        },
         meta: { errorMessage: "The update(s) could not be deleted" },
         onSuccess: async (data, variables) => {
             if (queryKey[0] === "profile") {
-                return queryClient.setQueryData(queryKey, (oldData: any) => ({
-                    ...oldData,
-                    userUpdates: [...oldData.userUpdates.filter((up: any) => up.id !== variables.updateIds[0]), data],
-                }));
+                return queryClient.setQueryData<ProfileOptionsType>(queryKey, (oldData) => {
+                    if (!oldData || !data) return;
+                    return {
+                        ...oldData,
+                        userUpdates: [...oldData.userUpdates.filter((up) => up.id !== variables.updateIds[0]), data],
+                    }
+                });
             }
             else if (queryKey[0] === "allUpdates") {
                 await queryClient.invalidateQueries({ queryKey });
             }
             else if (queryKey[0] === "onOpenHistory") {
-                return queryClient.setQueryData(queryKey, (oldData: any) => {
-                    return [...oldData.filter((history: any) => history.id !== variables.updateIds[0])];
+                return queryClient.setQueryData<HistoryOptionsType>(queryKey, (oldData) => {
+                    if (!oldData) return;
+                    return [...oldData.filter((history) => history.id !== variables.updateIds[0])];
                 });
             }
         },
@@ -40,20 +46,24 @@ export const useDeleteUpdatesMutation = (queryKey: DeleteUpdatesKeys) => {
 export const useAddMediaToListMutation = (mediaType: MediaType, mediaId: number | string, queryKey: DetailsUserListKeys) => {
     const queryClient = useQueryClient();
 
-    return useMutation<any, Error, { status?: Status }>({
-        mutationFn: ({ status }) => {
-            return postAddMediaToList({ data: { mediaId, mediaType, status } })
+    return useMutation({
+        mutationFn: ({ status }: { status?: Status }) => {
+            return postAddMediaToList({ data: { mediaId: mediaId, mediaType, status } })
         },
         meta: {
             successMessage: `${mediaType} added to your list`,
             errorMessage: `Failed to add this ${mediaType} to your list`,
         },
         onSuccess: (data) => {
-            queryClient.setQueryData(queryKey, (oldData: any) => {
-                if (queryKey[0] === "details") {
-                    return { ...oldData, userMedia: data };
-                }
-                else if (queryKey[0] === "userList") {
+            if (queryKey[0] === "details") {
+                queryClient.setQueryData<MediaDetailsOptionsType>(queryKey, (oldData) => {
+                    if (!oldData || !data) return;
+                    return { ...oldData, userMedia: data as UserMedia };
+                });
+            }
+            else if (queryKey[0] === "userList") {
+                queryClient.setQueryData<MediaListOptionsType>(queryKey, (oldData) => {
+                    if (!oldData) return;
                     return {
                         ...oldData,
                         results: {
@@ -63,8 +73,8 @@ export const useAddMediaToListMutation = (mediaType: MediaType, mediaId: number 
                             )
                         },
                     };
-                }
-            });
+                });
+            }
         }
     });
 };
@@ -73,25 +83,28 @@ export const useAddMediaToListMutation = (mediaType: MediaType, mediaId: number 
 export const useRemoveMediaFromListMutation = (mediaType: MediaType, mediaId: number | string, queryKey: DetailsUserListKeys) => {
     const queryClient = useQueryClient();
 
-    return useMutation<any, Error, {}>({
+    return useMutation({
         mutationFn: () => postRemoveMediaFromList({ data: { mediaId, mediaType } }),
         meta: { errorMessage: "Failed to remove this media from your list" },
         onSuccess: () => {
-            queryClient.setQueryData(queryKey, (oldData: any) => {
-                if (queryKey[0] === "details") {
-                    const { userMedia, ...rest } = oldData;
-                    return rest;
-                }
-                else if (queryKey[0] === "userList") {
+            if (queryKey[0] === "details") {
+                queryClient.setQueryData<MediaDetailsOptionsType>(queryKey, (oldData) => {
+                    if (!oldData) return;
+                    return { ...oldData, userMedia: null };
+                });
+            }
+            else if (queryKey[0] === "userList") {
+                queryClient.setQueryData<MediaListOptionsType>(queryKey, (oldData) => {
+                    if (!oldData) return;
                     return {
                         ...oldData,
                         results: {
                             ...oldData.results,
-                            items: [...oldData.results.items.filter((media: any) => media.mediaId !== mediaId)],
+                            items: [...oldData.results.items.filter((media) => media.mediaId !== mediaId)] as any,
                         },
                     };
-                }
-            });
+                })
+            }
         }
     });
 };
@@ -100,17 +113,21 @@ export const useRemoveMediaFromListMutation = (mediaType: MediaType, mediaId: nu
 export const useUpdateUserMediaMutation = (mediaType: MediaType, mediaId: number, queryKey: DetailsUserListKeys) => {
     const queryClient = useQueryClient();
 
-    return useMutation<Record<string, any>, Error, { payload: Record<string, any> }>({
-        mutationFn: ({ payload }) => {
+    return useMutation({
+        mutationFn: ({ payload }: { payload: Record<string, any> }) => {
             return postUpdateUserMedia({ data: { mediaType, mediaId, payload } });
         },
         meta: { errorMessage: "Failed to update this field value. Please try again later." },
         onSuccess: (data) => {
-            queryClient.setQueryData(queryKey, (oldData: Record<string, any>) => {
-                if (queryKey[0] === "details") {
-                    return { ...oldData, userMedia: { ...oldData.userMedia, ...data } };
-                }
-                else if (queryKey[0] === "userList") {
+            if (queryKey[0] === "details") {
+                queryClient.setQueryData<MediaDetailsOptionsType>(queryKey, (oldData) => {
+                    if (!oldData) return;
+                    return { ...oldData, userMedia: { ...oldData.userMedia, ...data as UserMedia } };
+                })
+            }
+            else if (queryKey[0] === "userList") {
+                queryClient.setQueryData<MediaListOptionsType>(queryKey, (oldData) => {
+                    if (!oldData) return;
                     return {
                         ...oldData,
                         results: {
@@ -120,8 +137,8 @@ export const useUpdateUserMediaMutation = (mediaType: MediaType, mediaId: number
                             }),
                         }
                     };
-                }
-            });
+                });
+            }
         },
     });
 };
@@ -130,21 +147,23 @@ export const useUpdateUserMediaMutation = (mediaType: MediaType, mediaId: number
 export const useEditUserLabelMutation = (mediaType: MediaType, mediaId: number) => {
     const queryClient = useQueryClient();
 
-    return useMutation<any, Error, { label: Label, action: LabelAction }>({
-        mutationFn: ({ label, action }) => {
+    return useMutation({
+        mutationFn: ({ label, action }: { label: Label, action: LabelAction }) => {
             return postEditUserLabel({ data: { mediaType, mediaId, label, action } });
         },
         meta: { errorMessage: "Failed to edit this label" },
         onSuccess: (data, variables) => {
-            queryClient.setQueryData(queryKeys.labelsKey(mediaType), (oldData: Label[]) => {
+            queryClient.setQueryData<(Label | undefined)[]>(queryKeys.labelsKey(mediaType), (oldData) => {
+                if (!oldData) return;
+
                 if (variables.action === "add") {
-                    return oldData.map(l => l.name).includes(data.name) ? oldData : [...oldData, data];
+                    return oldData.map(l => l?.name).includes(data?.name) ? oldData : [...oldData, data];
                 }
                 else if (variables.action === "rename") {
-                    return oldData.map(l => l.name === variables.label.oldName ? data : l);
+                    return oldData.map(l => l?.name === variables.label.oldName ? data : l);
                 }
                 else if (variables.action === "deleteAll") {
-                    return oldData.filter(l => l.name !== variables.label.name);
+                    return oldData.filter(l => l?.name !== variables.label.name);
                 }
             });
         }
